+++ src/isa/ebpf/inst.h
@@ -295,7 +295,7 @@ class ebpf_inst: public inst {
   int get_num_operands() const {return _isa.num_operands[_opcode];}
   int get_insn_num_regs() const {return _isa.insn_num_regs[_opcode];}
   int get_opcode_type() const {return _isa.opcode_type[_opcode];}
-  int64_t interpret(const vector<inst*> &instptr_list, prog_state &ps, int input = 0) const;
+  int64_t interpret(const vector<inst*> &instptr_list, prog_state &ps, int64_t input = 0) const;
   // smt
   z3::expr smt_inst(smt_var& sv) const;
   z3::expr smt_inst_jmp(smt_var& sv) const;
