+++ src/utils.h
@@ -16,7 +16,7 @@ using namespace std;
 #define L32(v) (0x00000000ffffffff & v)
 
 void print_test_res(bool res, string test_name);
-void gen_random_input(vector<int>& inputs, int min, int max);
+void gen_random_input(vector<int64_t>& inputs, int min, int max);
 ostream& operator<<(ostream& out, const vector<double>& vec);
 void split_string(const string& s, vector<string>& v, const string& c);
 unsigned int pop_count_asm(unsigned int x);
